[Feat/#85] 로그아웃 및 회원 탈퇴 API 구현 (#86)
* [Feat/#85] 로그아웃 API 구현

* [Feat/#85] 회원 탈퇴 API 구현

* [Feat/#85] JwtAuthenticationFilter 작성

* [Feat/#85] 커스텀 어노테이션 @AuthUser 작성

* [Feat/#85] @AuthUser 사용하여 API 인증 처리

* [Feat/#85] Swagger JWT Auth 설정

+++ src/main/java/com/friends/easybud/card/service/CardCommandServiceImpl.java
@@ -8,7 +8,6 @@ import com.friends.easybud.card.repository.CardRepository;
 import com.friends.easybud.global.exception.GeneralException;
 import com.friends.easybud.global.response.code.ErrorStatus;
 import com.friends.easybud.member.domain.Member;
-import com.friends.easybud.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,13 +17,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class CardCommandServiceImpl implements CardCommandService {
 
-    private final MemberRepository memberRepository;    // TODO MemberQueryService 주입
     private final CardRepository cardRepository;
 
     @Override
-    public Long createCard(CardCreateDto request) {
-        Member member = memberRepository.findById(1L).get();    // TODO 로그인 된 사용자 정보 조회
-
+    public Long createCard(Member member, CardCreateDto request) {
         Card card = buildCard(request, member);
         cardRepository.save(card);
 
@@ -42,9 +38,11 @@ public class CardCommandServiceImpl implements CardCommandService {
     }
 
     @Override
-    public Long updateCard(Long cardId, CardUpdateDto request) {
+    public Long updateCard(Member member, Long cardId, CardUpdateDto request) {
         Card card = cardRepository.findById(cardId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.CARD_NOT_FOUND));
+
+        checkCardOwnership(member, card);
 
         card.update(request);
 
@@ -52,12 +50,21 @@ public class CardCommandServiceImpl implements CardCommandService {
     }
 
     @Override
-    public Boolean deleteCard(Long cardId) {
+    public Boolean deleteCard(Member member, Long cardId) {
         Card card = cardRepository.findById(cardId)
                 .orElseThrow(() -> new GeneralException(ErrorStatus.CARD_NOT_FOUND));
+
+        checkCardOwnership(member, card);
+
         // TODO 연관된 Account 처리
         cardRepository.delete(card);
         return Boolean.TRUE;
+    }
+
+    private void checkCardOwnership(Member member, Card card) {
+        if (!card.getMember().equals(member)) {
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED_CARD_ACCESS);
+        }
     }
 
 }
